backend/requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=3.1.0
marshmallow>=3.0.0
Werkzeug>=3.0.0
bcrypt>=3.2.0
python-dotenv>=0.19.2
pytest>=7.1.2
requests>=2.26.0
email-validator>=2.0.0